Move TrackableActivity into common package
