+++ app/src/main/java/org/gdg/frisbee/android/common/GdgActivity.java
@@ -33,7 +33,6 @@ import com.google.android.gms.plus.Plus;
 
 import org.gdg.frisbee.android.R;
 import org.gdg.frisbee.android.achievements.AchievementActionHandler;
-import org.gdg.frisbee.android.activity.TrackableActivity;
 import org.gdg.frisbee.android.utils.PrefUtils;
 import org.gdg.frisbee.android.utils.ScopedBus;
 import org.gdg.frisbee.android.utils.Utils;
